Changed limit for social distancing from int to long

+++ src/main/kotlin/codeit/student/kotlintemplate/models/social/SocialDistancingResponse.kt
@@ -1,3 +1,3 @@
 package codeit.student.kotlintemplate.models.social
 
-data class SocialDistancingResponse (val answers: Map<String, Int>)
+data class SocialDistancingResponse (val answers: Map<String, Long>)
